+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.5.1] - 2026-09-21
+
 ### Added
 
 - Added the "Date format" setting for writing dates with tokens such as YYYY-MM-DD